backend/requirements.txt
# Python Dependencies for TTS Backend

# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Google Cloud
google-cloud-texttospeech==2.16.3
google-cloud-storage==2.14.0

# HTTP Client
aiohttp==3.9.1
httpx==0.26.0

# Environment
python-dotenv==1.0.0

# Security
python-multipart==0.0.6
python-jose[cryptography]==3.3.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
